requirements.txt
rich
python-dotenv